refactor(architecture): split dialogue system into workspace with specialized crates
This commit refactors the dialogue system architecture from a monolithic
crate into a proper Rust workspace with specialized subcrates:

- Create a workspace-level package "funkus_dialogue" that re-exports everything
- Extract core functionality into "funkus_dialogue_core" subcrate
- Move UI components into "funkus_dialogue_ui" subcrate
- Update examples to work with the new structure

These changes improve the project's maintainability by:
- Enforcing clearer separation of concerns between core and UI
- Allowing users to only import what they need
- Making dependencies more explicit
- Providing a cleaner path for future UI alternatives
- Creating a more idiomatic Rust package structure
