+++ crates/funkus_dialogue_core/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "funkus_dialogue_core"
+version = "0.1.0"
+edition = "2021"
+description = "Core functionality for the Bevy dialogue system"
+license = "Apache-2.0"
+repository = "https://github.com/StudioFunkus/funkus_dialogue"
+homepage = "https://github.com/StudioFunkus/funkus_dialogue"
+
+[dependencies]
+bevy = { version = "0.15", default-features = false }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = { version = "2", default-features = false }
+bevy_common_assets = { version = "0.12", features = ["json", "ron"] }
+petgraph = "0.7"
+bevy_egui = { version = "0.33", optional = true }
+
+[dev-dependencies]
+bevy = "0.15"
+
+[features]
+default = []
+debug_ui = ["bevy_egui"]
+
+[[example]]
+name = "simple_dialogue"
+path = "examples/simple_dialogue.rs"
