+++ crates/funkus_dialogue_core/src/asset/dialogue_asset.rs
