+++ crates/funkus_dialogue_core/src/asset/mod.rs
