+++ crates/funkus_dialogue_core/src/debug.rs
