+++ crates/funkus_dialogue_core/src/error.rs
