+++ crates/funkus_dialogue_core/src/events.rs
