+++ crates/funkus_dialogue_core/src/graph/node.rs
